requirements-pi.txt
-r requirements-dev.txt
# The following requirements are used by only the hardware modules for Raspberry Pi.
spidev >= 3.5
RPi.GPIO >= 0.7.1
RPi.bme280 >= 0.2.4
